requirements.txt
Flask>=1.0.0
Flask-SQLAlchemy>=2.0.0
click>=7.0.0
stringcase>=1.2.0